Working on making login as solid as possible

+++ web/backend/migrations/00000000000058_create_website_user_roles_table/up.sql
@@ -3,7 +3,7 @@
 CREATE TABLE website_user_roles (
     id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE cascade,
     website_role_id BIGINT NOT NULL REFERENCES website_roles(id) ON DELETE cascade,
-    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE cascade,
+    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE cascade,
     UNIQUE (website_role_id, user_id)
 );
